A shoe e-commerce website
This website is built with React.js, GSAP, Tailwind and Vite. It is a site that helps the user to be able to buy shoes from the supplier directly and also be able to buy in bulk from the supplier directly on a promo. it has good animations and very good interactive sections.